requirements.txt
argparse>=1.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0